BE/chrome_ux_report_project/requirements.txt
Django==4.2.6
djangorestframework==3.14.0
requests==2.31.0
django-cors-headers=4.3.0
